Initial module-aware changes for Go1.11
Changes required atop this fork of GopherJS for Go 1.11.

Includes an initial cut of module support, i.e. using GopherJS in a
module-aware context.

Not all tests are passing, but it's in good enough shape for a start to
merge into go1.11 branch.

Will need much more work to ensure that module-aware mode GopherJS works
just as well as GOPATH-mode GopherJS