build: calculate hash key based on augmented package files
Previously, we incorrectly determined cache staleness based on the
unaugmented files of a package. e.g. where there are special overrides
for various packages, these we not being taken into account. Now we do
take that into account, but at the slight cost of also unconditionally
parsing each file in the process (the code that augments a package is
rather tightly coupled to the parsing, so we leave that alone).

We also do not cache test packages either.