compiler: upgrade to Go 1.13
This is an initial cut of 1.13 support. internal/reflectlite is a new
package in Go 1.13 that presents a subset of the API of reflect.

internal/reflectlite support has been added according to the following
steps:

* in the build step, override the .GoFiles for internal/reflectlite to
be empty, i.e. we will only take the native (GopherJS) implementation
and won't augment any GOROOT definitions
* taking the current native (GopherJS) implementation of reflect
* taking the Go 1.13 implementation of reflect
* adding all files into natives/src/internal/reflectlite (marking the Go
1.13 reflect files as _original.go)
* progressively removing definitions from the *_original.go files until
we get back to a package that compiles. This involves removing certain
superfluous definitions (that only exist in the API of reflect) that
bring in unwanted imports, e.g. strconv).

We also then special case internal/reflectlite in the same places that
reflect is special-cased.

To handle the new core vendor of golang.org/x/net/dns/dnsmessage and
friends, we use .ImportMap from the output of go list to resolve the
correct package for any given import within a package.

WIP - we still need to fix the handling of import paths at test time.